kerilog/adminServer/requirements.txt
requests
paho-mqtt
enum
Flask
Flask-Admin
Flask-SQLAlchemy
Flask-Security>=1.7.5
MySQL-python>=1.2.5
